# perf: remove opaque_path distinction for storing