Before:
```
BasicBench_AdaURL            5144 ns         5141 ns       136020 time/byte=6.75509ns
```
After:
```
BasicBench_AdaURL            5020 ns         5016 ns       139829 time/byte=6.59072ns
```